Updating my fork of gitcoin/web

+++ app/marketing/views.py
@@ -24,9 +24,10 @@ import logging
 
 from django.conf import settings
 from django.contrib import messages
+from django.contrib.admin.views.decorators import staff_member_required
 from django.core.validators import validate_email
 from django.db.models import Max
-from django.http import Http404
+from django.http import Http404, HttpResponse
 from django.shortcuts import redirect
 from django.template.response import TemplateResponse
 from django.urls import reverse
@@ -35,6 +36,7 @@ from django.utils.translation import LANGUAGE_SESSION_KEY
 from django.utils.translation import gettext_lazy as _
 
 from app.utils import sync_profile
+from cacheops import cached_view
 from dashboard.models import Profile, TokenApproval
 from dashboard.utils import create_user_action
 from enssubdomain.models import ENSSubdomainRegistration
@@ -43,7 +45,7 @@ from mailchimp3 import MailChimp
 from marketing.mails import new_feedback
 from marketing.models import AccountDeletionRequest, EmailSubscriber, Keyword, LeaderboardRank
 from marketing.utils import get_or_save_email_subscriber, validate_discord_integration, validate_slack_integration
-from retail.emails import ALL_EMAILS
+from retail.emails import ALL_EMAILS, render_nth_day_email_campaign
 from retail.helpers import get_ip
 
 logger = logging.getLogger(__name__)
@@ -77,6 +79,9 @@ def get_settings_navs(request):
     }, {
         'body': _('Token'),
         'href': reverse('token_settings'),
+    }, {
+        'body': _('Job Status'),
+        'href': reverse('job_settings'),
     }]
 
 
@@ -107,7 +112,7 @@ def settings_helper_get_auth(request, key=None):
     # lazily create profile if needed
     profiles = Profile.objects.none()
     if github_handle:
-        profiles = Profile.objects.prefetch_related('alumni').filter(handle__iexact=github_handle).exclude(email='')
+        profiles = Profile.objects.prefetch_related('alumni').filter(handle__iexact=github_handle)
     profile = None if not profiles.exists() else profiles.first()
     if not profile and github_handle:
         profile = sync_profile(github_handle, user=request.user)
@@ -538,6 +543,90 @@ def account_settings(request):
                     subscriber_hash=subscriber_hash,
                 )
             except Exception as e:
+                logger.debug(e)
+            if es:
+                es.delete()
+            request.user.delete()
+            AccountDeletionRequest.objects.create(
+                handle=profile.handle,
+                profile={
+                        'ip': get_ip(request),
+                    }
+                )
+            profile.avatar_baseavatar_related.all().delete()
+            try:
+                profile.delete()
+            except:
+                profile.github_access_token = ''
+                profile.user = None
+                profile.hide_profile = True
+                profile.save()
+            messages.success(request, _('Your account has been deleted.'))
+            logout_redirect = redirect(reverse('logout') + '?next=/')
+            return logout_redirect
+        else:
+            msg = _('Error: did not understand your request')
+
+    context = {
+        'is_logged_in': is_logged_in,
+        'nav': 'internal',
+        'active': '/settings/account',
+        'title': _('Account Settings'),
+        'navs': get_settings_navs(request),
+        'es': es,
+        'profile': profile,
+        'msg': msg,
+    }
+    return TemplateResponse(request, 'settings/account.html', context)
+
+
+def job_settings(request):
+    """Display and save user's Account settings.
+
+    Returns:
+        TemplateResponse: The user's Account settings template response.
+
+    """
+    msg = ''
+    profile, es, user, is_logged_in = settings_helper_get_auth(request)
+
+    if not user or not profile or not is_logged_in:
+        login_redirect = redirect('/login/github?next=' + request.get_full_path())
+        return login_redirect
+
+    if request.POST:
+
+        if 'preferred_payout_address' in request.POST.keys():
+            profile.preferred_payout_address = request.POST.get('preferred_payout_address', '')
+            profile.save()
+            msg = _('Updated your Address')
+        elif request.POST.get('disconnect', False):
+            profile.github_access_token = ''
+            profile = record_form_submission(request, profile, 'account-disconnect')
+            profile.email = ''
+            profile.save()
+            create_user_action(profile.user, 'account_disconnected', request)
+            messages.success(request, _('Your account has been disconnected from Github'))
+            logout_redirect = redirect(reverse('logout') + '?next=/')
+            return logout_redirect
+        elif request.POST.get('delete', False):
+
+            # remove profile
+            profile.hide_profile = True
+            profile = record_form_submission(request, profile, 'account-delete')
+            profile.email = ''
+            profile.save()
+
+            # remove email
+            try:
+                client = MailChimp(mc_user=settings.MAILCHIMP_USER, mc_api=settings.MAILCHIMP_API_KEY)
+                result = client.search_members.get(query=es.email)
+                subscriber_hash = result['exact_matches']['members'][0]['id']
+                client.lists.members.delete(
+                    list_id=settings.MAILCHIMP_LIST_ID,
+                    subscriber_hash=subscriber_hash,
+                )
+            except Exception as e:
                 logger.exception(e)
             if es:
                 es.delete()
@@ -558,14 +647,14 @@ def account_settings(request):
     context = {
         'is_logged_in': is_logged_in,
         'nav': 'internal',
-        'active': '/settings/account',
-        'title': _('Account Settings'),
+        'active': '/settings/job',
+        'title': _('Job Settings'),
         'navs': get_settings_navs(request),
         'es': es,
         'profile': profile,
         'msg': msg,
     }
-    return TemplateResponse(request, 'settings/account.html', context)
+    return TemplateResponse(request, 'settings/job.html', context)
 
 
 def _leaderboard(request):
@@ -575,6 +664,9 @@ def _leaderboard(request):
         TemplateResponse: The leaderboard template response.
 
     """
+    context = {
+        'active': 'leaderboard',
+    }
     return leaderboard(request, '')
 
 
@@ -591,15 +683,19 @@ def leaderboard(request, key=''):
     if not key:
         key = 'quarterly_earners'
 
+    keyword_search = request.GET.get('keyword')
+    limit = request.GET.get('limit', 25)
+
     titles = {
         'quarterly_payers': _('Top Payers'),
         'quarterly_earners': _('Top Earners'),
         'quarterly_orgs': _('Top Orgs'),
         'quarterly_tokens': _('Top Tokens'),
         'quarterly_keywords': _('Top Keywords'),
-        # 'quarterly_cities': _('Top Cities'),
-        # 'quarterly_countries': _('Top Countries'),
-        # 'quarterly_continents': _('Top Continents'),
+        'quarterly_kudos': _('Top Kudos'),
+        'quarterly_cities': _('Top Cities'),
+        'quarterly_countries': _('Top Countries'),
+        'quarterly_continents': _('Top Continents'),
         #        'weekly_fulfilled': 'Weekly Leaderboard: Fulfilled Funded Issues',
         #        'weekly_all': 'Weekly Leaderboard: All Funded Issues',
         #        'monthly_fulfilled': 'Monthly Leaderboard',
@@ -621,14 +717,24 @@ def leaderboard(request, key=''):
         raise Http404
 
     title = titles[key]
-    leadeboardranks = LeaderboardRank.objects.active().filter(leaderboard=key)
-    amount = leadeboardranks.values_list('amount').annotate(Max('amount')).order_by('-amount')
-    items = leadeboardranks.order_by('-amount')
+    if keyword_search:
+        ranks = LeaderboardRank.objects.filter(active=True, leaderboard=key, tech_keywords__icontains=keyword_search)
+    else:
+        ranks = LeaderboardRank.objects.filter(active=True, leaderboard=key)
+
+    amount = ranks.values_list('amount').annotate(Max('amount')).order_by('-amount')
+    items = ranks.order_by('-amount')
+
     top_earners = ''
+    technologies = set()
+    for profile_keywords in ranks.values_list('tech_keywords'):
+        for techs in profile_keywords:
+            for tech in techs:
+                technologies.add(tech)
 
     if amount:
         amount_max = amount[0][0]
-        top_earners = leadeboardranks.order_by('-amount')[0:3].values_list('github_username', flat=True)
+        top_earners = ranks.order_by('-amount')[0:5].values_list('github_username', flat=True)
         top_earners = ['@' + username for username in top_earners]
         top_earners = f'The top earners of this period are {", ".join(top_earners)}'
     else:
@@ -636,8 +742,9 @@ def leaderboard(request, key=''):
 
     profile_keys = ['_tokens', '_keywords', '_cities', '_countries', '_continents']
     is_linked_to_profile = any(sub in key for sub in profile_keys)
+
     context = {
-        'items': items,
+        'items': items[0:limit],
         'titles': titles,
         'selected': title,
         'is_linked_to_profile': is_linked_to_profile,
@@ -646,6 +753,16 @@ def leaderboard(request, key=''):
         'card_desc': f'See the most valued members in the Gitcoin community recently . {top_earners}',
         'action_past_tense': 'Transacted' if 'submitted' in key else 'bountied',
         'amount_max': amount_max,
-        'podium_items': items[:3] if items else []
+        'podium_items': items[:5] if items else [],
+        'technologies': technologies,
+        'active': 'leaderboard'
     }
+
     return TemplateResponse(request, 'leaderboard.html', context)
+
+@staff_member_required
+def day_email_campaign(request, day):
+    if day not in list(range(1, 3)):
+        raise Http404
+    response_html, _, _, = render_nth_day_email_campaign('foo@bar.com', day, 'staff member')
+    return HttpResponse(response_html)
